Typecheck before running tests
Parcel doesn't check types by default (there is unstable support currently),
and relying on an IDE isn't useful when using the CLI.

This adds a #check npm script which does a basic typecheck with tsc,
and expands the #test npm script to run it before executing tests.

We also now run #check before a production #build or starting parcel
in dev server mode with #start).